io_uring: do not release write memory that the kernel still owns (#17238)
Motivation:

`shutdownOutput(...)` in `AbstractChannel` sets `this.outboundBuffer =
null` and then `closeOutboundBufferForShutdown(...)` releases the
flushed messages. In the io_uring transport a write SQE submitted
moments earlier is still in flight at that point, so the kernel keeps
reading from memory that was already returned to the allocator, and for
a pooled buffer possibly handed to someone else.

`close()` is not affected.
`AbstractIoUringChannel.AbstractUringUnsafe.close(...)` parks the
request in `delayedClose` and waits until `canCloseNow()` reports that
neither `WRITE_SCHEDULED` nor `READ_SCHEDULED` is set, so a completion
can never observe a dropped outbound buffer through that path.
`shutdownOutput()` gets no such protection.

What surfaced first was an NPE, because every write-completion path
dereferences `outboundBuffer()`. `IoUringIoHandler` catches `Throwable`
while dispatching a completion and logs it, so the channel never reaches
`ioState &= ~WRITE_SCHEDULED`, `canCloseNow()` stays false, the pending
`delayedClose` never runs, and the fd plus its `IoRegistration` leak. To
the application, `close()` simply never completes. The NPE and the leak
are symptoms; releasing memory the kernel still owns is the defect.

This is reachable without calling `shutdownOutput()` directly.
`handleWriteError(...)` routes every non-`IOException` write failure,
and every `IOException` when `AUTO_CLOSE` is disabled, into
`shutdownOutput(voidPromise(), t)`. epoll is unaffected because
`doWrite` there is synchronous, so there is no window in which a write
is outstanding while the outbound buffer is being dropped.

Resolves #17237

Modification:

Add `WriteOperation`, a slot recording the references a submitted write
handed to the kernel. Recording does not retain: the outbound buffer
still owns them. `doShutdownOutput()` retains what is in flight, and the
terminal CQE releases it, `IORING_CQE_F_NOTIF` for `IORING_OP_SEND_ZC`
and `IORING_OP_SENDMSG_ZC` and the completion itself for everything
else. `ChannelOutboundBuffer` and the shutdown path are untouched.

`WriteOperationTracker` owns the slots, one tracker per channel, so that
contract lives in one type. It holds four members: a pooled
`WriteOperation[]` indexed by the id in the SQE's user_data, a short
free stack that hands those ids out from 1 and takes them back on a
terminal CQE, an overflow map for the zero-copy ids that outlive the
short range, and the single slot a non-zero-copy stream write uses,
which needs no id because `WRITE_SCHEDULED` caps a stream channel at one
outstanding write. `retainAll()` and `releaseAll()` reach all four.

The datagram sendmsg path registers `MsgHdrMemoryArray` indices, which
that array allocates and recycles itself and which overlap with the ids
the free list hands out, so `recordForeign(...)` keeps them in their own
array.

Completions are matched on the opcode as well as the id, because write
`user_data` is not a single namespace. `IoUringFileRegion` submits fixed
values to tell its two splice stages apart, and the domain-socket
fd-passing sendmsg submits its `MsgHdrMemory` index. Neither is
registered, and with dense ids those values land inside the occupied
range rather than outside it, so the opcode is what keeps such a
completion from terminating what it finds.

`IovArrayReferenceCollector` captures exactly the buffers that entered
the iovec, including a composite that `IovArray` only added part of.
`IoUringIoHandler` owns it alongside the `IovArray` it fills, so one
instance serves every channel on an event loop.

`zcWriteQueue` and `ZC_BATCH_MARKER` are gone.
`handleWriteCompleteZeroCopy` no longer has to retain buffers itself, so
both the `IORING_CQE_F_MORE` and the non-`MORE` path use
`removeBytes(res)`, which also stops reporting a partial write as a full
one.

The outbound-buffer null checks stay in the five write-completion paths.
`shutdownOutput()` still nulls it, so a completion must not dereference
it even though the memory is now kept alive.

Result:

Write memory stays live until the kernel is done with it. `close()`
completes after `shutdownOutput()` races an in-flight write, and the fd
and its registration are released.

`IoUringShutdownOutputDuringInFlightWriteTest` covers the plain and the
zero-copy paths and asserts the buffer is still live right after
`shutdownOutput()`. It is not timing-dependent: the write SQE and
`shutdownOutput()` are issued in one event-loop task and
`submitAndRunNow()` is a no-op across the transport, so no CQE can have
been reaped by the time the outbound buffer is dropped.
`IoUringShutdownOutputDuringInFlightSpliceTest` covers the same race for
splice, where dropping the last reference to an `IoUringFileRegion`
closes both ends of its pipe and the `FileChannel` it splices from, all
of which the SQE carries as raw fd numbers.

`IoUringSocketSendSzSendmsgZcTest` covers the existing zero-copy paths
unchanged, with the reference-count assertions across
`IORING_CQE_F_MORE` and `IORING_CQE_F_NOTIF` tightened.
`IoUringDatagramWriteLifecycleTest` covers the sendmsg lifetime.
`IovArrayReferenceCollectorTest`, `WriteOperationTest`,
`WriteOperationTrackerTest` and `WriteOperationTrackerOverflowTest`
cover the collector, the release rules and all four tracker members on
the retain and release paths.

A JMH benchmark drives `WriteOperationTracker`, `WriteOperation` and
`IovArrayReferenceCollector` directly rather than a reimplementation of
them, so it cannot drift from what ships:

|                    | ops/s          | B/op |
|--------------------|----------------|------|
| stream single send | 150.8M ± 0.6M  | 0    |
| stream writev 4    | 21.5M ± 0.03M  | 0    |
| stream writev 16   | 8.2M ± 0.3M    | 0    |
| send_zc round trip | 93.8M ± 0.4M   | 0    |
| datagram sendmsg   | 119.5M ± 0.6M  | 0    |

64 byte pooled direct buffers, single thread, M1, 99.9% confidence.
Nothing allocates per write. Dropping the `arraycopy` in `record(byte,
ReferenceCounted[], int)` measures 27% at 4 iovecs and 6% at 16, but
that copy is what lets one collector be shared across an event loop, so
I kept it. The benchmark is not in this PR. I have it and can add it if
that is useful.

One gap I left open: the domain-socket fd-passing sendmsg is not
registered as a write operation, so it relies on `MsgHdrMemory.idx()`
being 0 and `nextOpsId()` never handing 0 out. That holds today and
nothing enforces it. I can register it as well if you would rather have
the invariant explicit.

---------

+++ transport-classes-io_uring/src/main/java/io/netty/channel/uring/AbstractIoUringServerChannel.java
@@ -147,7 +147,7 @@ abstract class AbstractIoUringServerChannel extends AbstractIoUringChannel imple
         }
 
         @Override
-        boolean writeComplete0(byte op, int res, int flags, short data, int outstanding) {
+        boolean writeComplete0(byte op, int res, int flags, long data, int outstanding) {
             throw new UnsupportedOperationException();
         }
